chore(sinks/sqs): implemented aws sqs
