--- src/sinks/_pending/aws_sqs/mod.rs
@@ -1,4 +1,0 @@
-mod run;
-mod setup;
-
-pub use setup::*;
